requirements.txt
fastapi
pydantic
uvicorn
pandas
psycopg2-binary